requirements.txt
streamlit
ollama
faster-whisper
ultralytics
ffmpeg-python
opencv-python-headless
pyyaml
torch
torchvision
torchaudio
